splitting views into smaller easier to understand views. Allow multiple images to be selected for place review
