+++ loc/Views/PlaceReviews/SelectedImagesView.swift
@@ -1,0 +1,35 @@
+//
+//  SelectedImagesView.swift
+//  loc
+//
+//  Created by Andrew Hartsfield II on 1/18/25.
+//
+
+
+import SwiftUI
+
+struct SelectedImagesView: View {
+    let images: [UIImage]
+
+    var body: some View {
+        if images.isEmpty {
+            Text("No images")
+                .foregroundColor(.gray)
+        } else {
+            // Horizontal scroll
+            ScrollView(.horizontal, showsIndicators: false) {
+                HStack(spacing: 10) {
+                    ForEach(images, id: \.self) { image in
+                        Image(uiImage: image)
+                            .resizable()
+                            .scaledToFill()
+                            .frame(width: 100, height: 100)
+                            .clipShape(RoundedRectangle(cornerRadius: 10))
+                    }
+                }
+                .padding(.horizontal)
+            }
+        }
+    }
+}
+
